Refactored axum resources
