+++ central/Cargo.toml
@@ -15,19 +15,21 @@ uuid = { version = "1.1.2", features = [
     "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
     "serde"
 ]}
-url = "2.2.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 axum = { version = "0.5.9", features = [ "query" ] }
 #axum-macros = "0.2.2"
 ctrlc = { version = "3.2.2", features = ["termination"] }
-static_init = "1.0.2"
 
 anyhow = "*"
+thiserror = "1.0.31"
 
 # Logging is imported through shared
 tracing = "0.1.35"
 #tracing-subscriber = { version = "0.3.11", features = ["env-filter"] }
+
+hyper = { version = "0.14.19", features = ["full"] }
+hyper-tls = "0.5.0"
 
 [build-dependencies]
 build-data = "0"
