+++ proxy/Cargo.toml
@@ -11,8 +11,6 @@ shared = { path = "../shared", features = ["config-for-proxy"] }
 tokio = { version = "1", features = ["full"] }
 async-recursion = "1.0.0"
 
-static_init = "1.0.2"
-
 # Web Server & clean shutdown
 axum = "0.5.9"
 hyper = { version = "0.14.19", features = ["full"] }
